Changed the way it checks if current number is a multiple of prime factors
Before there used to be an array MULT, however this way it is clear that
we are checking if the current number is a multiple of prime factors to
date. Also, there was a bug due to incorrect indexing in the original
code. The modulus seems to solve most of it however a few changes need
to still be made in order to print 2 as a prime factor.